Extract reusable amrap_partial_shares helper and amrap_round_sizes_for
Splits the AMRAP partial-round math out of assign_amrap_set_breakdown_targets
into standalone helpers (amrap_partial_shares, amrap_round_sizes_for) so the
round-size reconstruction can be reused for display grouping, not just target
validation. Moves the AMRAP round-sizing methods into their own concern
(LogAmrapRoundSizing) to keep LogScoring under the module length limit.
